pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ape-zksync"
version = "0.1.0"
description = "zkSync Era plugin for Ape Framework"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "eth-ape>=0.8,<0.9",
    "zksync2>=2.0.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-ape",
]

[project.entry-points."ape_cli_subcommands"]
zksync = "ape_zksync.__main__:cli"

[project.entry-points."ape_plugins"]
zksync = "ape_zksync"
